Separated email template file and functions
