+++ src/myFunctions/sendEmailVerification.ts
@@ -1,16 +1,10 @@
 import ejs from 'ejs';
-import formData from 'form-data';
-import Mailgun from 'mailgun.js';
-import emailTemplate from './emailTemplate';
+import emailTemplateEmailVerification from './emailTemplateEmailVerification';
 import uuid from 'uuid';
 import disposableEmails from '../util/disposableEmails.js';
+import { sendEmail } from './sendEmail';
 
-const api_key = process.env.MAILGUN_API_KEY;
-const domain = process.env.PROAVALON_EMAIL_ADDRESS_DOMAIN;
 const server_domain = process.env.SERVER_DOMAIN;
-
-const mailgun = new Mailgun(formData);
-const mg = mailgun.client({ username: 'api', key: api_key });
 
 const uuidv4 = uuid.v4;
 
@@ -27,20 +21,17 @@ export const sendEmailVerification = (user: any, email: string) => {
 
   const token = uuidv4();
 
-  const message = ejs.render(emailTemplate, { server_domain, token });
-
-  const data = {
-    from: 'ProAvalon <' + process.env.PROAVALON_EMAIL_ADDRESS + '>',
-    to: user.emailAddress,
-    subject: 'Welcome! Please verify your email address.',
-    html: message,
-  };
-
   user.emailToken = token;
   user.markModified('emailToken');
   user.save();
 
-  mg.messages.create(domain, data);
+  const message = ejs.render(emailTemplateEmailVerification, {
+    server_domain,
+    token,
+  });
+  const subject = 'Welcome! Please verify your email address.';
+
+  sendEmail(email, subject, message);
 };
 
 export const isThrowawayEmail = (email: string): boolean => {
